requirements.txt
flickrapi
Pillow
cliff
